Fixed multiple warnings: conversion from 'size_t' to int/unsigned

+++ unittest/spmc_queue.cpp
@@ -160,7 +160,7 @@ void wsq_test_n_thieves(int N) {
     auto num_stolen = [&] () {
       int total = 0;
       for(const auto& cdeq : cdeqs) {
-        total += cdeq.size();
+        total += static_cast<int>(cdeq.size());
       }
       return total;
     };
